@@ -13,11 +13,19 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^4.0|^5.0"
     },
     "autoload": {
         "psr-4": {
-            "Endroid\\Bundle\\ImagePlaceholderBundle\\": [ "src/", "tests/" ]
+            "Endroid\\Bundle\\ImagePlaceholderBundle\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Endroid\\Bundle\\ImagePlaceholderBundle\\Tests\\": "tests/"
         }
     },
     "config": {
