@@ -14,7 +14,7 @@ class BeerHoldProvider implements ProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getUrl($width, $height, array $options = array())
+    public function getUrl($width, $height, array $options = [])
     {
         $url = 'http://beerhold.it/'.$width.'/'.$height;
 
