@@ -14,7 +14,7 @@ class PlaceCreatureProvider implements ProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getUrl($width, $height, array $options = array())
+    public function getUrl($width, $height, array $options = [])
     {
         $url = 'http://placecreature.com/'.$width.'/'.$height;
 
