@@ -14,7 +14,7 @@ class PlaceHoffProvider implements ProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getUrl($width, $height, array $options = array())
+    public function getUrl($width, $height, array $options = [])
     {
         $url = 'http://place-hoff.com/'.$width.'/'.$height;
 
